Fix a bug with upgrading all modules
Prior to this change, the code to collect all modules was picking up zsh
extensions, which are not accesible as modules in their own right.

This change collects all modules and then checks the config to see if
they're installed. This means the only modules collected will be modules
that can be accessed directly.