Pipe the output from subprocess calls to rich
Prior to this change, the output from subprocesses was being piped
directly to stdout/stderr. This will cause issues with statuses in rich.

This change collects stdout and stderr from subprocesses and sends them
to the terminal via rich.

NOTE: this puts all output on stdout, which appears to be unavoidable.
This probably doesn't matter, though.