Dynamically determine zshrc file path when needed
Prior to this change, we stored the zshrc filepath for a module in
__init__. This meant that the path was calculated at module
instantiation time, which is also import time. This meant the config
file was being loaded at import time, which caused errors when setting
up a new machine (since the config file does not yet exist)

This change moves the calculation of the zshrc filepath to the method
that needs it: there's no reason to store it on the object.