Set the PAGER and use it for diff-so-fancy
Prior to this change, the default PAGER was less, and we hard-coded less
as the pager for diffs with diff-so-fancy.

This change sets the default PAGER to be less with the -RFX flags. If
less has been upgraded, we remove the -X flag because it breaks
scrolling. If bat has been installed then we use that.