Show tidier output when printing exceptions
Prior to this change, we printed the full traceback in a glaring red
when there was an exception installing or upgrading a module. This is
messy and confusing and our application deliberately returns useful
messages so is probably unhelpful.

This change just prints the error message, which should be sufficient.