Only install production dependencies
Prior to this change, we installed dev dependencies when we updated qaz.
We deliberately don't install these during the initial install so
there's no reason we should here.

This change prevents the dev dependencies being installed on update.

Nite that this will cause them to be uninstalled once this code is live.