[Starship] Use fill instead of right prompt
Prior to this change, we showed some items on the right of the prompt
using the 'right_format' option. However, this put those items in the
right prompt, which is on the same line as the input (after everything
else).

This change uses the `fill` module to shift the same items to the right
but in the main prompt line and befroethe linebreak and other modules on
the same line as the character.