Make `clb` alias show commits
Prior to this change, the `clb` alias cleared the screen and printed the
branch name in big fancy letters at the top. I never use this.

This change replaces it with a command to clear the screen and then
print the log from the main brnch until the current head. I have this
set up as a post-checkout hook at work and will use it much more often.