[VS Code] add some rules to better-comments config
Prior to this change, we had this extension installed to do some comment
highlighting but weren't using it for 'important' notes (because the
convention at work is not to use '!' or '*' to denote important
comments).

This change adds 'N.B.' and 'note' as important comments.

We also have to add in the default rules so we don't lose them.