[VS Code] Remove better comments config
Prior to this change, we added a rule to highlight comments starting
with 'note' or 'Note'. However, this meant all lines starting with
'note' were highlighted, even if they weren't comments. This was very
annoying (particularly where that is a common argument/variable name).

This change removes that config and restores the default highlighting
rules.