Require Python to be installed before QAZ
Prior to this change, we installed ASDF and Python as part of setting up
QAZ. However, this was a lot for the install script to do and I don't
want to require ASDF for installing Python (there are better ways).

This change assumes Python is already installed. This is a reasonable
requirement.