Fix import formatting
Prior to this change, pre-commit started messing up the import sorting
with isort. It appears to have been picking up the configuration from
`configfiles/.editorconfig` instead of `setup.cfg`.

It appears that this is related to PyCQA/isort#1872 and
PyCQA/isort#1907.

This change explicitly sets the settings file to ensure the right file
is used.

We also fix the incorrect imports.