[VS Code] Replace GitHub extension
Prior to this change, we used a GitHub extension that has been
deprecated in favour of GH's official PR/Issue management extension.
However, that does not provide a way to open the current file in GH,
which is theo nly part of the extension I use.

This change installs a different extension, which allows easy opening
of files. We also configure it to point to the remote HEAD instead of
the current branch.